Fix game logic for checking guesses
Update function to check guesses only after compiling all guesses
into the 'guess' object. Do this by using map to map the individual
pitches to strings and store those in a new array. Fix definition of
'checkTreble' variable so that the original 'treble' variable in the
global scope is not changed. Update 'includes' ('yellow-overlay') guess
checking to remove the specific note that is referenced by 'includes'.